Re-organize files. Create MLIR backend classes
