+++ src/contrib/mlir/CMakeLists.txt
@@ -14,13 +14,96 @@
 # limitations under the License.
 # ******************************************************************************
 
-# Common MLIR/LLVM setup for the sub-directories below
-
 include_directories(
     ${NGRAPH_INCLUDE_PATH}
     ${MLIR_LLVM_INCLUDE_PATH}
     ${MLIR_INCLUDE_PATHS}
 )
 
-add_subdirectory(compiler)
 add_subdirectory(tools/ngraph-opt)
+
+set(SRC
+    backend/cpu_backend.cpp
+    backend/pass/affine_lowerer.cpp
+    backend/pass/memory_optimization.cpp
+    core/compiler.cpp
+    core/dialect/dialect.cpp
+    core/dialect/type.cpp
+    core/dialect/ops.cpp
+    core/pass/mlir_subgraph_extraction.cpp
+    core/pass/mlir_subgraph_extraction.hpp
+    core/utils.cpp
+    runtime/cpu/memory_manager.cpp
+)
+
+add_library(mlir_backend SHARED ${SRC})
+
+llvm_map_components_to_libnames(llvm_libs support core irreader)
+
+# Link MLIR libs
+target_link_libraries(
+    mlir_backend PRIVATE 
+    MLIRAnalysis
+    MLIREDSC
+    MLIRExecutionEngine
+    MLIRIR
+    MLIRLLVMIR
+    MLIRStandardToLLVM
+    MLIRParser
+    MLIRPass
+    MLIRTargetLLVMIR
+    MLIRTransforms
+    MLIRSupport
+)
+# some libs need whole archive linkage because of Globals static initialization
+function(whole_archive_link target)
+    if("${CMAKE_SYSTEM_NAME}" STREQUAL "Darwin")
+        set(link_flags "-Llib -Wl,-all_load ")
+        FOREACH(LIB ${ARGN})
+            string(CONCAT link_flags ${link_flags} "${LIB}")
+        ENDFOREACH(LIB)
+    else()
+        set(link_flags "-Llib -Wl,--whole-archive,")
+        FOREACH(LIB ${ARGN})
+            string(CONCAT link_flags ${link_flags} "${LIB},")
+        ENDFOREACH(LIB)
+        string(CONCAT link_flags ${link_flags} "--no-whole-archive")
+    endif()
+    message(STATUS "MLIR Ops link flag: ${link_flags}" )
+
+    set_target_properties(${target} PROPERTIES LINK_FLAGS ${link_flags})
+endfunction(whole_archive_link)
+
+whole_archive_link(mlir_backend
+    ${LLVM_BUILD_LIBRARY_DIR}/libMLIRAffineOps.a
+    ${LLVM_BUILD_LIBRARY_DIR}/libMLIRStandardOps.a
+)
+# Link LLVM libs
+target_link_libraries(
+    mlir_backend PRIVATE 
+    ${llvm_libs}
+)
+
+# Link ngraph 
+target_link_libraries(mlir_backend PUBLIC ngraph)
+
+# table-gen dialect ops
+# include table-gen helpers
+include(${LLVM_DIR}/TableGen.cmake)
+
+function(ngraph_tablegen ofn)
+    tablegen(MLIR ${ARGV} "-I${MLIR_SRC_INCLUDE_PATH}" "-I${MLIR_BIN_INCLUDE_PATH}")
+    set(TABLEGEN_OUTPUT ${TABLEGEN_OUTPUT} ${CMAKE_CURRENT_BINARY_DIR}/${ofn} PARENT_SCOPE)
+endfunction()
+
+set(MLIR_TABLEGEN_EXE mlir-tblgen)
+set(LLVM_TARGET_DEFINITIONS core/dialect/ops.td)
+
+ngraph_tablegen(ops.h.inc -gen-op-decls)
+ngraph_tablegen(ops.cpp.inc -gen-op-defs)
+
+add_public_tablegen_target(ngraph_ops_gen)
+add_dependencies(mlir_backend ngraph_ops_gen)
+target_include_directories(mlir_backend PRIVATE ${CMAKE_CURRENT_BINARY_DIR})
+
+install(TARGETS mlir_backend DESTINATION ${NGRAPH_INSTALL_LIB})
