+++ src/ngraph/runtime/cpu/cpu_external_function.cpp
@@ -36,7 +36,7 @@
 #endif
 
 #ifdef NGRAPH_MLIR_ENABLE
-#include "contrib/mlir/compiler/pass/mlir_subgraph_extraction.hpp"
+#include "contrib/mlir/core/pass/mlir_subgraph_extraction.hpp"
 #endif
 
 #include "ngraph/descriptor/input.hpp"
