+++ src/ngraph/runtime/cpu/cpu_runtime_context.hpp
@@ -31,7 +31,7 @@
 #include "ngraph/op/experimental/compiled_kernel.hpp"
 
 #ifdef NGRAPH_MLIR_ENABLE
-#include "contrib/mlir/compiler/compiler.hpp"
+#include "contrib/mlir/core/compiler.hpp"
 #endif
 
 namespace mkldnn
